Update file structure for ScenarioBlob
Created a new directory called ScenarioBlob to organize all the files related to ScenarioBlob together.

Also added unit tests for ScenarioBlob.
