--- src/PaperG/FirehoundBlob/BlobInterface.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace PaperG\FirehoundBlob;
-
-
-interface BlobInterface {
-    /**
-     * @return array
-     */
-    public function toArray();
-
-    /**
-     * @param $array array
-     */
-    public function fromArray($array);
-}
