+++ src/PaperG/ScenarioBlob/BlobInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace PaperG\ScenarioBlob;
+
+
+interface BlobInterface {
+    /**
+     * @return array
+     */
+    public function toArray();
+
+    /**
+     * @param $array array
+     */
+    public function fromArray($array);
+}
